Add feedback when adding or deleting recipes

+++ src/components/RecipeCard.jsx
@@ -1,22 +1,29 @@
 import { Card, Button, Row, Col } from "react-bootstrap";
-import { useContext } from "react";
+import { useContext, useState } from "react";
 import DataContext from "../contexts/DataContext";
 
 const RecipeCard = (props) => {
 
     const [recipes, setRecipes] = useContext(DataContext);
 
+    const [deleteable, setDeletable] = useState(false);
+
     const deleteRecipe = ()=>{
-        fetch(`https://324t9imtmd.execute-api.us-east-2.amazonaws.com/default/recipes/${props.recipeID}`, {
-            method: "DELETE"
-        }).then(resp => {
-            if(resp.status === 200){
-                setRecipes(recipes.filter((r) => r.recipeID !== props.recipeID))
-            }
-            else{
-                alert("Could not delete");
-            }
-        })
+        if (!deleteable) {
+            setDeletable(true);
+        }
+        else {
+            fetch(`https://324t9imtmd.execute-api.us-east-2.amazonaws.com/default/recipes/${props.recipeID}`, {
+                method: "DELETE"
+            }).then(resp => {
+                if(resp.status === 200){
+                    setRecipes(recipes.filter((r) => r.recipeID !== props.recipeID))
+                }
+                else{
+                    alert("Could not delete");
+                }
+            })
+        }
     }
 
 
@@ -28,7 +35,7 @@ const RecipeCard = (props) => {
                         <h1>{props.recipeName}</h1>
                     </Col>
                     <Col xs sm md lg xl = {2} style={{display:'flex', justifyContent:'right'}}>
-                        <Button onClick = {deleteRecipe} style={{backgroundColor:"red", borderColor:"black"}}>Delete Recipe</Button>
+                        <Button onClick = {deleteRecipe} style={{backgroundColor:"red", borderColor:"black"}}>{deleteable ? 'Are you sure?' : 'Delete Recipe'}</Button>
                     </Col>
                 </Row>
                 <h3>Ingredients</h3>
